refactor: clean up start models
